Modify hstore adaptor to only produce strings
The prior implementation would adapt the types, and the following would happen:

input:  {'foo': True, 'bar': "baz"}
output: hstore(ARRAY['foo', 'bar'], ARRAY[true, 'baz'])

The output value of foo is a boolean type, which is not acceptable in an hstore (at least not with this creation method).

Anyways, coercing everything to a string solves the issue. I am using the unicode function to make the tests happy.