Officially deprecate psycopg1 module
Per the module's comment, using psycopg1 is considered a temporary hack
that should not be relied on long term. Import now reports a warning.
The warning instructs users to use the modern module instead. Allows
removing the workaround in a future version.

The psycopg1 has been available since the very first psycopg2 module.
Allowing plenty of time for users to migrate.