Fix definition of BUILD_DIR in Makefile
The Python's platform should be included in the build dir. Was
previously missing resulting in the incorrect path.

Fixing the build dir allows testing psycopg2 without first calling
"python setup.py install".

Exposed a missing use of the deprecated "async" keyword.

For additional information on how to calculate the build dir, see the
following SO post:

https://stackoverflow.com/a/14369968