Take libdir from pg_config --ldflags
On OS X with OpenSSL installed via brew, `-lssl` is not found
during the linker stage. This can be remedied by correctly taking into
account the build-time settings for PostgreSQL.

Since the format is incompatible, this commit introduces some string
parsing to retrieve the directories referenced in the linker call.