Emit a warning when a connection or cursor isn't closed
This follows the semantics of Python file objects. When the object is
garbage collected and it has not been closed, emit a ResourceWarning.
This allows library users to notice inconsistent and non-deterministic
resource management and fix it. Users should use context managers and
finally blocks to always close cursors and connections when they are not
longer required.

For example, in Python, not closing a file results in the following:

    $ python3 -Walways
    >>> f = open('foo', 'w')
    >>> del f
    __main__:1: ResourceWarning: unclosed file <_io.TextIOWrapper name='foo' mode='w' encoding='UTF-8'>
    ResourceWarning: Enable tracemalloc to get the object allocation traceback

psycopg now acts the same way:

    $ python3 -Walways
    >>> import psycopg2
    >>> c = psycopg2.connect(database='psycopg2_test')
    >>> del c
    <stdin>:1: ResourceWarning: unclosed connection <connection object at 0x7fc4aec38d50; dsn: 'dbname=psycopg2_test', closed: 0>
    ResourceWarning: Enable tracemalloc to get the object allocation traceback

All warnings noticed during testing has been fixed.