Remove unnecessary sensationalism in docs
SQL injection and failures are bad but this seems over the top. In a
literal sense I'm struggling to imagine a scenario where bad
interpolation is worth getting shot over. And I understand this was
probably intended as hyperbole, but I don't think there's any need to
invoke the spectre of violence to make the point here.