Fixes #1252 - Increment refcount on provided parameters.
When parsing an object (`O`) with `PyArg_ParseTuple` and friends, the
provided object is set in the appropiate field, but the refcount is not
incremented. Reference:
https://docs.python.org/3/c-api/arg.html#other-objects.

Use `Py_XINCREF` to increment the refcount on any objects that may have
been passed into this function. `Py_XINCREF` performs a NULL check
internally and only increments the refcounter on non-NULL objects.